+++ pipeline.build.ps1
@@ -62,7 +62,7 @@ function CopyModuleFiles {
     }
 }
 
-task VersionModule PSRule, {
+task VersionModule ModuleDependencies, {
     $modulePath = Join-Path -Path $ArtifactPath -ChildPath PSRule.Rules.Azure;
     $manifestPath = Join-Path -Path $modulePath -ChildPath PSRule.Rules.Azure.psd1;
     Write-Verbose -Message "[VersionModule] -- Checking module path: $modulePath";
@@ -103,11 +103,9 @@ task VersionModule PSRule, {
             Update-ModuleManifest -Path $manifestPath -Prerelease $revision;
         }
     }
-
-    $manifest = Get-Content -Path $manifestPath -Raw;
-    $manifest.Replace('RequiredModules = @()', "RequiredModules = @(@{ ModuleName = 'PSRule'; ModuleVersion = '0.5.0' }, @{ ModuleName = 'Az.Accounts'; ModuleVersion = '1.4.0' }, @{ ModuleName = 'Az.StorageSync'; ModuleVersion = '0.8.0' }, @{ ModuleName = 'Az.Security'; ModuleVersion = '0.7.4' }, @{ ModuleName = 'Az.Storage'; ModuleVersion = '1.1.1' }, @{ ModuleName = 'Az.Websites'; ModuleVersion = '1.1.2' }, @{ ModuleName = 'Az.Sql'; ModuleVersion = '1.7.0' })") | Set-Content -Path $manifestPath;
 }
 
+# Synopsis: Publish to PowerShell Gallery
 task ReleaseModule VersionModule, {
     $modulePath = (Join-Path -Path $ArtifactPath -ChildPath PSRule.Rules.Azure);
     Write-Verbose -Message "[ReleaseModule] -- Checking module path: $modulePath";
@@ -116,7 +114,6 @@ task ReleaseModule VersionModule, {
         Write-Error -Message "[ReleaseModule] -- Module path does not exist";
     }
     elseif (![String]::IsNullOrEmpty($NuGetApiKey)) {
-        # Publish to PowerShell Gallery
         Publish-Module -Path $modulePath -NuGetApiKey $NuGetApiKey;
     }
 }
@@ -168,6 +165,28 @@ task platyPS {
     Import-Module -Name PlatyPS -Verbose:$False;
 }
 
+# Synopsis: Install module dependencies
+task ModuleDependencies NuGet, PSRule, {
+    if ($Null -eq (Get-InstalledModule -Name Az.Accounts -ErrorAction Ignore)) {
+        Install-Module -Name Az.Accounts -Scope CurrentUser -Force;
+    }
+    if ($Null -eq (Get-InstalledModule -Name Az.Resources -ErrorAction Ignore)) {
+        Install-Module -Name Az.Resources -Scope CurrentUser -Force;
+    }
+    if ($Null -eq (Get-InstalledModule -Name Az.Storage -ErrorAction Ignore)) {
+        Install-Module -Name Az.Storage -Scope CurrentUser -Force -AllowClobber;
+    }
+    if ($Null -eq (Get-InstalledModule -Name Az.Security -ErrorAction Ignore)) {
+        Install-Module -Name Az.Security -Scope CurrentUser -Force;
+    }
+    if ($Null -eq (Get-InstalledModule -Name Az.Sql -ErrorAction Ignore)) {
+        Install-Module -Name Az.Sql -Scope CurrentUser -Force;
+    }
+    if ($Null -eq (Get-InstalledModule -Name Az.Websites -ErrorAction Ignore)) {
+        Install-Module -Name Az.Websites -Scope CurrentUser -Force;
+    }
+}
+
 task CopyModule {
     CopyModuleFiles -Path src/PSRule.Rules.Azure -DestinationPath out/modules/PSRule.Rules.Azure;
 
@@ -182,6 +201,11 @@ task BuildModule CopyModule
 task TestRules PSRule, Pester, PSScriptAnalyzer, {
     # Run Pester tests
     $pesterParams = @{ Path = $PWD; OutputFile = 'reports/pester-unit.xml'; OutputFormat = 'NUnitXml'; PesterOption = @{ IncludeVSCodeMarker = $True }; PassThru = $True; };
+
+    if ($CodeCoverage) {
+        $pesterParams.Add('CodeCoverage', (Join-Path -Path $PWD -ChildPath 'out/modules/**/*.psm1'));
+        $pesterParams.Add('CodeCoverageOutputFile', (Join-Path -Path $PWD -ChildPath reports/pester-coverage.xml));
+    }
 
     if (!(Test-Path -Path reports)) {
         $Null = New-Item -Path reports -ItemType Directory -Force;